backend/requirements.txt
# FastAPI & Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0

# Supabase
supabase>=2.3.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.24.0

# Social Platform SDKs
atproto>=0.0.46
tweepy>=4.14.0

# AI
openai>=1.12.0
anthropic>=0.18.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
email-validator>=2.0.0
redis>=5.0.0
celery>=5.3.0
Pillow>=10.2.0

# Testing
pytest>=7.0.0,<8.0.0
pytest-asyncio>=0.23.0
